Add draft support for basketball
Implements draft support for basketball. Copied over from the football implementation. Only change is the string representation in order to better display the round and pick number.

This should probably be moved to a base pick class, similar to what is planned for players with #124 , but this should do for now. Might also be worth considering adding a player object member instead of the separate name and id-- similar to how the team member is a team object